+++ state_machine/state_machine_SPI.h
@@ -38,7 +38,7 @@ void init_state_machine_spi(Spi_State_Machine_t);
 void event_handler_spi(Spi_State_Machine_t spi_type);
 void requestSpiTransmit(Spi_State_Machine_t spi_type, uint8_t child_id, 
     uint16_t packet, uint32_t* read_var_addr);
-void requestSpiTransmit_remote(event_type_t msg_type, uint8_t data, uint8_t* read_var_addr);
+void requestSpiTransmit_remote(event_type_t msg_type, uint8_t data, uint32_t* read_var_addr);
 
 
 #ifdef __cplusplus
